requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
joblib==1.3.2
pandas==2.2.1
scikit-learn==1.4.0
requests==2.31.0
